server/service_whisper/requirements_whisper.txt
accelerate==1.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.9
aiosignal==1.3.1
alembic==1.14.0
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.6.2.post1
appdirs==1.4.4
asteroid-filterbanks==0.4.0
async-timeout==5.0.1
asyncio==3.4.3
attrs==24.2.0
audioread==3.0.0
bitsandbytes==0.45.0
blinker==1.9.0
cachetools==5.5.0
cffi==1.15.1
click==8.1.8
coloredlogs==15.0.1
colorlog==6.9.0
contourpy==1.3.1
cycler==0.12.1
decorator==5.1.1
distro==1.9.0
dnspython==2.7.0
docopt==0.6.2
efficientnet_pytorch==0.7.1
einops==0.8.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.8
fastapi-cli==0.0.7
ffmpeg-python==0.2.0
flatbuffers==24.12.23
fonttools==4.55.3
frozenlist==1.5.0
future==1.0.0
git-filter-repo==2.47.0
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.1.1
groq==0.13.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.0
huggingface-hub==0.27.0
humanfriendly==10.0
HyperPyYAML==1.2.2
inquirerpy==0.3.4
joblib==1.3.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
julius==0.2.7
kiwisolver==1.4.8
lazy_loader==0.3
librosa==0.10.2.post1
lightning==2.5.0.post0
llamaapi==0.1.36
llvmlite==0.40.1
Mako==1.3.8
markdown-it-py==3.0.0
matplotlib==3.10.0
mdurl==0.1.2
more-itertools==10.5.0
msgpack==1.0.5
multidict==6.1.0
narwhals==1.20.1
natsort==8.4.0
nest-asyncio==1.6.0
numba==0.57.1
omegaconf==2.3.0
onnxruntime==1.20.1
opencv-python==4.7.0.72
optuna==4.1.0
pandas==2.0.3
pfzy==0.3.4
pooch==1.6.0
primePy==1.3
prompt_toolkit==3.0.48
propcache==0.2.1
protobuf==5.29.2
psutil==6.1.1
pyannote.audio==3.3.2
pyannote.core==5.0.0
pyannote.database==5.1.0
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pyarrow==18.1.0
pycparser==2.21
pydantic==2.10.3
pydantic_core==2.27.1
pydeck==0.9.1
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.2.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.20
pytorch-lightning==2.5.0.post0
pytorch-metric-learning==2.8.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.11.6
rich==13.9.4
rich-toolkit==0.13.2
rpds-py==0.22.3
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
safetensors==0.4.5
scikit-learn==1.3.0
scipy==1.14.1
semver==3.0.2
sentencepiece==0.2.0
shellingham==1.5.4
silero-vad==5.1.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.12.1
soxr==0.3.5
speechbrain @ git+https://github.com/speechbrain/speechbrain.git@5403c777c44c4b9734ca548ca7008bb946229975
SQLAlchemy==2.0.36
starlette==0.45.3
streamlit==1.41.1
tabulate==0.9.0
tenacity==9.0.0
tensorboardX==2.6.2.2
threadpoolctl==3.1.0
tiktoken==0.8.0
tokenizers==0.21.0
toml==0.10.2
torch-audiomentations==0.11.2
torch_pitch_shift==1.2.5
tornado==6.4.2
transformers==4.47.1
typer==0.15.1
typing_extensions==4.12.2
tzdata==2023.3
uvicorn==0.34.0
uvloop==0.21.0
vad==1.0.2
watchdog==6.0.0
watchfiles==1.0.4
wcwidth==0.2.13
websockets==15.0
whisper-openai==1.0.0
yarl==1.18.3